+++ NEWS.md
@@ -5,13 +5,17 @@ All notable Changes to the Julia package `LieGroups.jl` will be documented in th
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.1.2] 2025-06-23
+## [0.1.2] 2025-06-24
 
 ### Added
 
 * `is_flat` for `SpecialEuclideanGroup`
 * `inner` and `norm` for `LieAlgebra` to compute the inner product and norm on the Lie algebra.
 * a test suite function for `identity_element`.
+* New StaticArrays.jl specializations for multiple functions, including:
+  * `exp` and `log` on the orthogonal and special orthogonal group in 2 and 3 dimensions.
+  * `get_coordinates` and `get_vector` on the orthogonal and special orthogonal group in 2 and 3 dimensions, for `LieAlgebraOrthogonalBasis`.
+* More generic implementation of non-mutating `get_vector_lie` on `AbstractProductGroupOperation` groups.
 
 ### Changed